% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{contrast_ratio}
\alias{contrast_ratio}
\title{WCAG 2.x contrast ratio between two sRGB colours}
\usage{
contrast_ratio(rgb1, rgb2)
}
\arguments{
\item{rgb1, rgb2}{Numeric vectors of length 3, sRGB channels in [0, 1].}
}
\value{
Contrast ratio between 1 and 21.
}
\description{
WCAG 2.x contrast ratio between two sRGB colours
}
