% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{hex_to_oklch}
\alias{hex_to_oklch}
\title{Convert a HEX colour to OKLCH}
\usage{
hex_to_oklch(hex_str)
}
\arguments{
\item{hex_str}{Colour as HEX string.}
}
\value{
Numeric vector c(L, C, H): OKLab lightness, chroma, hue in degrees.
}
\description{
Convert a HEX colour to OKLCH
}
