% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{oklch_to_hex}
\alias{oklch_to_hex}
\title{Convert an OKLCH colour to HEX}
\usage{
oklch_to_hex(lch)
}
\arguments{
\item{lch}{Numeric vector c(L, C, H).}
}
\value{
HEX string (sRGB, channel values clipped to [0, 1]).
}
\description{
Convert an OKLCH colour to HEX
}
